Add files via upload
This version currently detects network scans like Nmap, pings & SSH attempts on port 22 it can be avoided otherwise, also it can detect any data transfers that involve machines that were not present when you initially launched the script. This is a very rough draft but I welcome any suggestions or contributions. The goal is to offer something simple that doesn't involve the heavy lifting involved with network security. Just run it on your raspberry pi or any machine and get simple alerts and logs. I am currently trying to tweak the output to the terminal to limit the amount of alerts. I also plan to research a program to help with Intrusion Detection.